print controller fixed
